input stored in state
Input in all the fields updates the state and calculate button logs state to console. The app now needs to be able to send the state values as an array to flask back and fetch the computed results to output a storage reccomendation. The app.js needs conditional rendering of pages.